added session loading state, refactor loading states into loading file components
